creating api.js for better text editor ergonomics
