+++ frontend/src/api/index.js
@@ -1,27 +1,1 @@
-function checkStatus(response) {
-  if (response.status >= 200 && response.status < 300) {
-    return response;
-  }
-  const error = new Error(response.statusText);
-  error.response = response;
-  throw error;
-}
-
-function parseJSON(response) {
-  return response.json();
-}
-
-export default {
-  async post(route, body) {
-    const response = await fetch(`api${route}`, {
-      method: 'POST',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(body),
-    });
-    const responseChecked = await checkStatus(response);
-    return parseJSON(responseChecked);
-  },
-};
+export { default } from './api';
